rename azure->dataloader to better reflect functionality
