+++ ehr2meds/PREMEDS/preprocessing/io/data_handling.py
@@ -5,7 +5,7 @@ from typing import Iterator, Optional
 import pandas as pd
 
 from ehr2meds.PREMEDS.preprocessing.constants import FILENAME
-from ehr2meds.PREMEDS.preprocessing.io.azure import get_data_loader
+from ehr2meds.PREMEDS.preprocessing.io.dataloader import get_data_loader
 
 logger = logging.getLogger(__name__)
